[PHP] run tests with PHP 8
PHP 7.x is not supported anymore (end of support of PHP 7.4 was in
November 2022). The latest version is PHP 8.2, however I guess it makes
sense to test with the oldest still maintained version, to make sure we
don't use featurase not supported by those

Note that I only change:
-the version used in CI (in order to ensure our sdk is compatible with
 supported version of PHP)
- the version mentioned in README (to avoid giving the feeling that our
  repo is obsolete)
but that I don't change the requirement in the generated composer file
on purpose, in order to not take the risk to break compatibility for
clients with old PHP version.