Update PHP composer packages
For some reason we're using a custom template to generate the
composer.json file instead of using the one provided by
openApiGenerator.

With c5c7ef7d720b73b5456ce67cceaf4223eb58b531 we upgraded
openApiGenerator (from 5.2.1 to 6.3.0), but we did not upgraded our
composer.mustache in the process.

FTR: the composer template used by openApiGenerator 6.3.0 is available at
https://github.com/OpenAPITools/openapi-generator/blob/v6.3.0/modules/openapi-generator/src/main/resources/php/composer.mustache
and declares those dependencies

```
"php": "^7.4 || ^8.0",
"ext-curl": "*",
"ext-json": "*",
"ext-mbstring": "*",
"guzzlehttp/guzzle": "^7.3",
"guzzlehttp/psr7": "^1.7 || ^2.0"
```

so the main discrepency with what we generate is the constraint on the
version of guzzlehttp/guzzle (that we constrained to be < 7.0.0)

Though the generated code seems to work rather fine, it still means that
on the first hand we have generated code designed to work with guzzle
7.3+ but that we constrained another version at runtime.

This also fixes
- https://github.com/criteo/criteo-api-marketingsolutions-php-sdk/issues/6
- and https://github.com/criteo/criteo-api-marketingsolutions-php-sdk/issues/1