Fix PHP example (and subsequent test)
ExampleApplication was introduced by c0c482d90eaa8c411ea969630546cdb1d6973669.
Unfortunately this filed had a typo (fixed by this current commit).

This commit also added a test, in GatewayApiTest, to run this example
code. However, this commit also changed the destination of that
generated test file, from `test/Api` to `test`.

But we run only tests located in `test/Api` (because of the config in
our generated composer.json), which means that since this commit we
stopped running test from GatewayApiTest at all.
That explains why this typo was not caught by tests in the first place.

What this current commit does is:
* fix the typo in the example file
* fix the location of the generated GatewayApiTest so those tests are run
* make some required adjustment in the generated test file (namely:
  importing the file with the example code, and adjusting the namespace
  to be able to reference the example class)

This fixes
https://github.com/criteo/criteo-api-marketingsolutions-php-sdk/issues/3