don't ignore jq error
the error we observed for CMS could have been easier to spot if the script
failed along with the jq command. But bash eats the intermediate error code
in a pipeline.
This change ensures that if we have again in the future some unexpected OAS
structure, then the action will fail instead of pushing a corrupted update